Add expand hook to fancytree in tree.js
Selected child elements beyond the second level in the tree are currently not selected. By adding the expand hook and checking if the childrens keys of the expanded node are present in the selected array on the ajaxData object we can set the child node as selected.
However i did not find a solution yet when deselecting the child nodes and toggling the expand of the parent. Then the items are reselected.